+++ libraries/template-haskell/changelog.md
@@ -10,6 +10,8 @@
 
   * Add support for unboxed sums. (#12478)
 
+  * Add support for visible type applications. (#12530)
+
 ## 2.11.0.0  *May 2016*
 
   * Bundled with GHC 8.0.1